requirements.txt
# Minimal requirements
requests
python-dotenv
Pillow
python-osc
pyyaml
pyttsx3
flask
dxcam
# Windows Graphics Capture
windows-capture
